requirements.txt
pandas==1.3.5
langchain-community==0.0.37
langchain==0.1.17
python-docx==0.8.11
groq==0.5.0
python-dotenv==0.19.1
playwright==1.43.0
html2text==2024.2.26
#streamlit==1.34.0
gradio==4.29.0
